Revert "feat(tibuild): add EventID, direct Tekton trigger, TektonReconciler, …"
This reverts commit b70aab718bd5198fbf50345111c2b6247967dba0.
