+++ tibuild/pkg/rest/service/model.go
@@ -115,7 +115,6 @@ type DevBuildStatus struct {
 }
 
 type TektonStatus struct {
-	EventID   string           `json:"eventID,omitempty"`
 	Pipelines []TektonPipeline `json:"pipelines"`
 }
